Move VolatileCell to libcommon
Fixes Issue #33.
