+++ src/chips/nrf51822/lib.rs
@@ -1,7 +1,9 @@
 #![crate_name = "nrf51822"]
 #![crate_type = "rlib"]
-#![feature(asm,core_intrinsics,concat_idents,const_fn)]
+#![feature(asm,concat_idents,const_fn)]
 #![no_std]
 
+extern crate common;
 extern crate hil;
+
 pub mod gpio;
